Consensus boundary identification

Consensus boundary identification

ConsensusTADs(
  cont_mats,
  resolution,
  z_thresh = 3,
  window_size = 15,
  gap_thresh = 0.2
)

Arguments

cont_mats
List of contact matrices in either sparse 3 column, n x n or n x (n+3) form where the first three columns are coordinates in BED format. See "Input_Data" vignette for more information. If an x n matrix is used, the column names must correspond to the start point of the corresponding bin. Required.

resolution
Resolution of the data. Used to assign TAD boundaries to genomic regions. If not provided, resolution will be estimated from column names of the first matrix. Default is "auto"

z_thresh
Threshold for boundary score. Higher values result in a higher threshold for differential TADs. Default is 3.

window_size
Size of sliding window for TAD detection, measured in bins. Results should be consistent Default is 15.

gap_thresh
Required % of non-zero entries before a region will be considered non-informative and excluded. Default is .2

Value

A list containing consensus TAD boundaries and overall scores

Consensus - Data frame containing location of all consensus boundaries. Coordinate is the region of the genome, Sample columns correspond to individual boundary scores. Consensus_Score is consensus boundary score.

All_Regions - Data frame containing consensus scores for all regions. All columns are identiical to the Consensus object.

Details

Given a list of sparse 3 column, n x n , or n x (n+3) contact matrices, ConsensusTADs provides the set of consensus TAD boundaries across them. Consensus TADs are defined by the consensus boundary score, a score measuring TAD boundary likelihood across all matrices.

Examples

# Read in data
data("time_mats")
# Find consensus TAD boundaries
diff_list <- ConsensusTADs(time_mats, resolution = 50000)

#> Converting to n x n matrix

#> Matrix dimensions: 704x704

#> Matrix dimensions: 704x704

#> Matrix dimensions: 704x704

#> Matrix dimensions: 704x704